@@ -3,7 +3,7 @@ description: Demonstrates how to use the flutter_video_compress plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -11,9 +11,9 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  image_picker: ^0.6.0+10
-  video_player: ^0.10.1+3
+  cupertino_icons: ^1.0.4
+  image_picker: ^0.8.4+4
+  video_player: ^2.2.6
   
 dev_dependencies:
   flutter_test:
