@@ -5,7 +5,7 @@ author: Tenkai Ruri<babichan@outlook.com>
 homepage: https://github.com/TenkaiRuri/flutter_video_compress
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -14,8 +14,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  image_picker: ^0.6.0+10
-  video_player: ^0.10.1+3
+  image_picker: ^0.8.4+4
+  video_player: ^2.2.6
 
 
 # For information on the generic Dart part of this file, see the
