@@ -1,24 +1,24 @@
 part of flutter_video_compress;
 
 class MediaInfo {
-  String path;
-  String title;
-  String author;
-  int width;
-  int height;
+  String? path;
+  String? title;
+  String? author;
+  int? width;
+  int? height;
 
   /// [Android] API level 17
-  int orientation;
+  int? orientation;
 
   /// bytes
-  int filesize; // filesize
+  int? filesize; // filesize
   /// microsecond
-  double duration;
-  bool isCancel;
-  File file;
+  double? duration;
+  bool? isCancel;
+  File? file;
 
   MediaInfo({
-    @required this.path,
+    this.path,
     this.title,
     this.author,
     this.width,
@@ -40,7 +40,7 @@ class MediaInfo {
     filesize = json['filesize'];
     duration = double.tryParse('${json['duration']}');
     isCancel = json['isCancel'];
-    file = File(path);
+    file = File(path ?? "");
   }
 
   Map<String, dynamic> toJson() {
@@ -58,7 +58,7 @@ class MediaInfo {
     if (this.isCancel != null) {
       data['isCancel'] = this.isCancel;
     }
-    data['file'] = File(path).toString();
+    data['file'] = File(path ?? "").toString();
     return data;
   }
 }
