@@ -12,7 +12,7 @@ class ObservableBuilder<T> {
   }
 
   Subscription subscribe(void onData(T event),
-      {Function onError, void onDone(), bool cancelOnError}) {
+      {Function? onError, void onDone()?, bool? cancelOnError}) {
     _notSubscribed = false;
     _observable.stream.listen(onData,
         onError: onError, onDone: onDone, cancelOnError: cancelOnError);
